backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.4
python-dotenv==1.0.1
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
prophet==1.1.5
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
httpx==0.27.0
openai==1.35.0
pdfplumber==0.11.0
